+++ src/ps/plone/mls/browser/developments/details.py
@@ -14,6 +14,7 @@ import logging
 from Acquisition import aq_inner
 from Products.CMFPlone import PloneMessageFactory as PMF
 from Products.Five import BrowserView
+from Products.Five.browser.pagetemplatefile import ViewPageTemplateFile
 from plone import api as plone_api
 from plone.app.layout.viewlets.common import ViewletBase
 from plone.directives import form
@@ -58,6 +59,8 @@ except ImportError:
 # local imports
 from ps.plone.mls import (
     _,
+    PLONE_4,
+    PLONE_5,
     api,
     config,
     utils,
@@ -99,11 +102,11 @@ function loadScript(src, callback) {{
 
 
 function loadGoogleMaps(callback) {{
-    if (typeof google === 'object' && typeof google.maps === 'object') {{
-        callback();
-    }} else {{
-        loadScript('https://maps.googleapis.com/maps/api/js?key={apikey}', callback);
-    }}
+  if (typeof google === 'object' && typeof google.maps === 'object') {{
+    callback();
+  }} else {{
+    loadScript('https://maps.googleapis.com/maps/api/js?key={ak}', callback);
+  }}
 }}
 
 
@@ -343,9 +346,24 @@ class DevelopmentDetails(BrowserView):
     _contact_form = None
     _contact_info = None
 
-    def __init__(self, context, request):
-        super(DevelopmentDetails, self).__init__(context, request)
+    if PLONE_5:
+        index = ViewPageTemplateFile('templates/development_details_p5.pt')
+    elif PLONE_4:
+        index = ViewPageTemplateFile('templates/development_details_view.pt')
+
+    def __call__(self):
+        self.setup()
+        return self.render()
+
+    def render(self):
+        return self.index()
+
+    def setup(self):
         self.registry = getUtility(IRegistry)  # noqa
+        if PLONE_5:
+            from Products.CMFPlone.resources import add_resource_on_request
+            if self.use_fotorama():
+                add_resource_on_request(self.request, 'psplonefotorama')
 
     @property
     def config(self):
@@ -427,7 +445,7 @@ class DevelopmentDetails(BrowserView):
             lng=lng,
             map_id=self.map_id,
             zoom=self.config.get('map_zoom_level', 7),
-            apikey=self.googleapi,
+            ak=self.googleapi,
         )
 
     @property